Move AuthProvider to Auth folder
