+++ src/GoogleCloudTracer/Auth/AuthProvider.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace HelloFresh\GoogleCloudTracer;
+namespace HelloFresh\GoogleCloudTracer\Auth;
 
 use Doctrine\Common\Cache\ArrayCache;
 use Doctrine\Common\Cache\Cache;
